Extras

Scats.Internal.Extras — Module

Module containing extras: auxiliary functions and macros used within the package.

Base.println — Method

println(io::IO, array::Array{RT, 1})

Print an array containing RT values.

Usage

using Scats.Internal.Prec: RT
using Scats.Internal.Extras: println

x = Array{RT}([1.0, 2.0, 3.0])
println(x)

# output

 1.000000000000000E+00    2.000000000000000E+00    3.000000000000000E+00

Note: this doctest requires x64 architecture.

Base.println — Method

println(io::IO, value::RT)

Print an RT value.

Usage

using Scats.Internal.Prec: RT
using Scats.Internal.Extras: println

x = RT(1.0)
println(x)

# output

 1.000000000000000E+00

Note: this doctest requires x64 architecture.

Scats.Internal.Extras.@exception — Macro

@exception(
    docstring::AbstractString,
    module_name::AbstractString,
    exception_name::AbstractString,
    error_message::AbstractString,
)

Construct an exception.

Usage

using Scats.Internal.Extras: @exception
@exception(
    "Exception thrown when the file is not found.",
    "Gen",
    "ScatsGenNotAFile",
    "The file is not found (\"", e.file, "\").",
)

This macro call is equivalent to this exception definition:

"""
    ScatsGenNotAFile <: Exception

Exception thrown when the file is not found.

"""
mutable struct ScatsGenNotAFile <: Exception
    file::AbstractString
    ScatsGenNotAFile(file::AbstractString) = new(file)
end

Base.showerror(io::IO, e::ScatsGenNotAFile) =
print(
    io, string(
        "\n\n",
        "Scats.Internal.Gen.ScatsGenNotAFile:\n",
        "The file is not found (\"", e.file, "\").\n"
    )
)

Scats.Internal.Extras.@pack — Macro

@pack(
    struct_instance::AbstractString,
    fields::AbstractString...,
)

Pack the specified variables in the structure instance fields of the same name.

Usage

using Scats.Internal.Extras: @pack
using Scats.Internal.Input: InputStruct
Input = InputStruct()

N = 230
Δt = 1
q = 0.01

@pack("Input", "N", "Δt", "q") ===
begin
Input.N = N
Input.Δt = Δt
Input.q = q
end

Scats.Internal.Extras.@unpack — Macro

@unpack(
    struct_instance::AbstractString,
    fields::AbstractString...,
)

Unpack an instance of a structure on the specified fields.

Usage

using Scats.Internal.Extras: @unpack
using Scats.Internal.Gen: GenStruct
Gen = GenStruct()

@unpack("Gen", "N", "Δt", "q", "α", "β", "r", "A", "ν", "ϕ", "γ") ===
begin
N = Gen.N
Δt = Gen.Δt
q = Gen.q
α = Gen.α
β = Gen.β
r = Gen.r
A = Gen.A
ν = Gen.ν
ϕ = Gen.ϕ
γ = Gen.γ
end