Gen

Scats.Internal.Gen — Module

Module containing the type to interact with the time series generator.

Scats.Internal.Gen.GenStruct — Type

GenStruct()

Instantiate this type to interact with the time series generator.

Data

N::IT=0: sample size;

Δt::RT=0: sample step;

q::RT=0: significance level;

α::RT=0: parameter α of a linear trend;

β::RT=0: parameter β of a linear trend;

r::IT=0: number of harmonics;

A::Vector{RT}=[]: amplitudes array;

ν::Vector{RT}=[]: frequencies array;

ϕ::Vector{RT}=[]: phase shifts array;

γ::RT=0: «signal-to-noise» ratio.

Methods

read!(file::AbstractString): read generator parameters from a file;

example(file::AbstractString): generate an example of a file containing the generator parameters;

gen!(): generate time series;

reset!(): reset an instance to default values.

Note

Data can be also read when calling an instance like so:

using Scats
s = Scats.API()
file, _ = mktemp()
s.Gen.example(file)
s.Gen(file)

Scats.Internal.Gen.gen! — Method

gen!(Gen::GenStruct, Input::InputStruct)

Generate time series for an instance of InputStruct using generator parameters from an instance of GenStruct.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Gen.example(file)
s.gen!()

Scats.Internal.Gen.ScatsGenEOF — Type

ScatsGenEOF <: Exception

Exception thrown when an end of file occurred.

Scats.Internal.Gen.ScatsGenIsADir — Type

ScatsGenIsADir <: Exception

Exception thrown when the path is a directory.

Scats.Internal.Gen.ScatsGenNotAFile — Type

ScatsGenNotAFile <: Exception

Exception thrown when the file is not found.

Scats.Internal.Gen.ScatsGenWR_A — Type

ScatsGenWR_A <: Exception

Exception thrown when an input for A is wrong.

Scats.Internal.Gen.ScatsGenWR_N — Type

ScatsGenWR_N <: Exception

Exception thrown when an input for N is wrong.

Scats.Internal.Gen.ScatsGenWR_q — Type

ScatsGenWR_q <: Exception

Exception thrown when an input for q is wrong.

Scats.Internal.Gen.ScatsGenWR_r — Type

ScatsGenWR_r <: Exception

Exception thrown when an input for r is wrong.

Scats.Internal.Gen.ScatsGenWR_Δt — Type

ScatsGenWR_Δt <: Exception

Exception thrown when an input for Δt is wrong.

Scats.Internal.Gen.ScatsGenWR_α — Type

ScatsGenWR_α <: Exception

Exception thrown when an input for α is wrong.

Scats.Internal.Gen.ScatsGenWR_β — Type

ScatsGenWR_β <: Exception

Exception thrown when an input for β is wrong.

Scats.Internal.Gen.ScatsGenWR_γ — Type

ScatsGenWR_γ <: Exception

Exception thrown when an input for γ is wrong.

Scats.Internal.Gen.ScatsGenWR_ν — Type

ScatsGenWR_ν <: Exception

Exception thrown when an input for ν is wrong.

Scats.Internal.Gen.ScatsGenWR_ϕ — Type

ScatsGenWR_ϕ <: Exception

Exception thrown when an input for ϕ is wrong.

Scats.Internal.Gen.example — Method

example(file::AbstractString)

Generate an example of a file containing the generator parameters.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Gen.example(file)

Scats.Internal.Gen.read! — Method

read!(Gen::GenStruct, file::AbstractString)

Read generator parameters from a file to an instance of GenStruct.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Gen.example(file)
s.Gen.read!(file)

Scats.Internal.Gen.reset! — Method

reset!(Gen::GenStruct)

Reset an instance of GenStruct to default values.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Gen.example(file)
s.Gen.read!(file)
s.Gen.reset!()

Scats.Internal.Gen.@array — Macro

@array(array::AbstractString)

Wrap an array for zero-based indexing.

Usage

using OffsetArrays
using Scats.Internal.Input: InputStruct
using Scats.Internal.Gen: @array

file, _ = mktemp()
Input = InputStruct()

Input.example(file)
Input(file)

N₋₁ = Input.N - 1
@array("Input.t") === OffsetArray(Input.t, 0:N₋₁)

Scats.Internal.Gen.@read — Macro

@read(element::AbstractString, type::AbstractString)

Read a scalar element.

Usage

using Scats.Internal.Gen: @read

@read("N", "IT") #= ===
begin
    try
        Gen.N = parse(IT, split(readline(io))[1])
    catch
        throw(ScatsGenWR_N(file))
    end
end =#

Scats.Internal.Gen.@read_array — Macro

@read_array(element::AbstractString, type::AbstractString)

Read an array element.

Usage

using Scats.Internal.Gen: @read_array

@read_array("A", "RT") #= ===
begin
    try
        Gen.A = parse.(RT, split(readline(io))[1:Gen.r])
    catch
        throw(ScatsGenWR_A(file))
    end
end =#

Scats.Internal.Gen.@skip — Macro

@skip()

Skip two lines and check for EOF each time.

Usage

using Scats.Internal.Gen: @skip

@skip() #= ===
begin
    readline(io)
    if eof(io)
        throw(ScatsGenEOF(file))
    end
    readline(io)
    if eof(io)
        throw(ScatsGenEOF(file))
    end
end =#