Input

Scats.Internal.Input — Module

Module containing the type to interact with input data.

Scats.Internal.Input.InputStruct — Type

InputStruct()

Instantiate this type to interact with input data.

Data

N::IT=0: sample size;

Δt::RT=0: sample step;

q::RT=0: significance level;

t::Vector{RT}=[]: time array;

x::Vector{RT}=[]: values array.

Methods

read!(file::AbstractString): read input data from a file;

write(file::AbstractString): write input data to a file;

example(file::AbstractString): generate an example of the input/output file;

reset!(): reset an instance to default values.

Note

Data can be also read when calling an instance like so:

using Scats
s = Scats.API()
file, _ = mktemp()
s.Input.example(file)
s.Input(file)

Scats.Internal.Input.ScatsInputEOF — Type

ScatsInputEOF <: Exception

Exception thrown when an end of file occurred.

Scats.Internal.Input.ScatsInputIsADir — Type

ScatsInputIsADir <: Exception

Exception thrown when the path is a directory.

Scats.Internal.Input.ScatsInputNotAFile — Type

ScatsInputNotAFile <: Exception

Exception thrown when the file is not found.

Scats.Internal.Input.ScatsInputWR_N — Type

ScatsInputWR_N <: Exception

Exception thrown when an input for N is wrong.

Scats.Internal.Input.ScatsInputWR_q — Type

ScatsInputWR_q <: Exception

Exception thrown when an input for q is wrong.

Scats.Internal.Input.ScatsInputWR_t — Type

ScatsInputWR_t <: Exception

Exception thrown when an input for t is wrong.

Scats.Internal.Input.ScatsInputWR_x — Type

ScatsInputWR_x <: Exception

Exception thrown when an input for x is wrong.

Scats.Internal.Input.ScatsInputWR_Δt — Type

ScatsInputWR_Δt <: Exception

Exception thrown when an input for Δt is wrong.

Scats.Internal.Input.example — Method

example(file::AbstractString)

Generate an example of the input/output file.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Input.example(file)

See: InputStruct.

Scats.Internal.Input.read! — Method

read!(Input::InputStruct, file::AbstractString)

Read input data from a file to an instance of InputStruct.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Input.example(file)
s.Input.read!(file)

Scats.Internal.Input.reset! — Method

reset!(input::InputStruct)

Reset an instance of InputStruct to default values.

Usage

using Scats
s = Scats.API()
s.Input.reset!()

Scats.Internal.Input.write — Method

write(input::InputStruct, file::AbstractString)

Write input data from an instance of InputStruct to a file.

Usage

using Scats
s = Scats.API()
file, _ = mktemp()
s.Input.write(file)

Scats.Internal.Input.@read — Macro

@read(element::AbstractString, type::AbstractString)

Read a scalar element.

Usage

using Scats.Internal.Input: @read

@read("N", "IT") #= ===
begin
    try
        Input.N = parse(IT, split(readline(io))[1])
    catch
        throw(ScatsInputWR_N(file))
    end
end =#

Scats.Internal.Input.@read_array — Macro

@read_array(element::AbstractString, type::AbstractString)

Read an array element.

Usage

using Scats.Internal.Input: @read_array

@read_array("t", "RT") #= ===
begin
    try
        Input.t = parse.(RT, split(readline(io))[1:Input.N])
    catch
        throw(ScatsInputWR_t(file))
    end
end =#

Scats.Internal.Input.@skip — Macro

@skip()

Skip two lines and check for EOF each time.

Usage

using Scats.Internal.Input: @skip

@skip() #= ===
begin
    readline(io)
    if eof(io)
        throw(ScatsInputEOF(file))
    end
    readline(io)
    if eof(io)
        throw(ScatsInputEOF(file))
    end
end =#