Module

All inner parts of this package are conveniently hidden in the inner module Scats.Internal:

Scats.Internal — Module

A module containing all inner parts of this package. Is not advisable to be used directly.

See documentation for more info about its contents.

The next section presents the module containing the precisions and the formats of numbers used in the package.

The following sections present the main types used inside the package which are appearing in an API instance:

Input: input data;

Gen: generator;

Result: result data.

The last section shows what extras are developed for the usage inside the package.