Public

Main Module

Scats — Module

A package for completing spectral correlation analysis of time series.

Important note: this package was developed primarily for
learning purposes and is not advised to be used in production.

Links:
- Repo: https://github.com/paveloom-j/Scats.jl
- Docs: https://paveloom-j.github.io/Scats.jl

Please, use the published docs instead of retrieving
information from these docstrings manually.

Interface

This part of the package is the only one supposed to be used by a regular user. All recommended for usage types and methods are accessible from an instance of Scats.API. For the inner parts one should consider reading the pages in the Internals section.

Scats.API — Type

API()

Instantiate an instance of Scats API to get access to the public interface.

Types

Input: input data;

Gen: generator;

Result: result data.

Methods

for Input:

read_input!(file::AbstractString): read input data from a file;

write_input(file::AbstractString): write input data to a file;

input_example(file::AbstractString): generate an example of the input/output file.

for Gen:

read_gen!(file::AbstractString): read generator parameters from a file;

gen_example(file::AbstractString): generate an example of a file containing the generator parameters;

gen!(): generate time series.

Usage

using Scats
s = Scats.API()