Gen

Scats.internal.gen — Module

Module containing a type for interaction with the time series generator.

Scats.internal.gen.GenStruct — Type

GenStruct()

Instantiate this type to interact with the time series generator.

Data

N::IT=0: sample size;

Δt::RT=0: sample step;

q::RT=0: significance level;

α::RT=0: parameter α of a linear trend;

β::RT=0: parameter β of a linear trend;

r::IT=0: number of harmonics;

A::Vector{RT}=[]: amplitudes array;

ν::Vector{RT}=[]: frequencies array;

ϕ::Vector{RT}=[]: phase shifts array;

γ::RT=0: «signal-to-noise» ratio.

Methods

read!(file::AbstractString): read generator parameters from a file;

example(file::AbstractString): generate an example of a file containing the generator parameters;

gen(): generate time series;

reset!(): reset an instance to default values.

Note

Data can be also read calling an instance like so:

using Scats
s = Scats.api()
file, _ = mktemp()
s.gen.example(file)
s.gen(file)

Scats.internal.gen.gen! — Method

gen!(gen::GenStruct, input::InputStruct)

Generate time series for an instance of InputStruct using generator parameters from an instance of GenStruct.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.gen.example(file)
s.gen!()

Scats.internal.gen.ScatsGenEOF — Type

ScatsGenEOF <: Exception

Exception thrown when an end of file occurred.

Scats.internal.gen.ScatsGenIsADir — Type

ScatsGenIsADir <: Exception

Exception thrown when the path is a directory.

Scats.internal.gen.ScatsGenNotAFile — Type

ScatsGenNotAFile <: Exception

Exception thrown when the file is not found.

Scats.internal.gen.ScatsGenWR_A — Type

ScatsGenWR_A <: Exception

Exception thrown when an input for A is wrong.

Scats.internal.gen.ScatsGenWR_N — Type

ScatsGenWR_N <: Exception

Exception thrown when an input for N is wrong.

Scats.internal.gen.ScatsGenWR_q — Type

ScatsGenWR_q <: Exception

Exception thrown when an input for q is wrong.

Scats.internal.gen.ScatsGenWR_r — Type

ScatsGenWR_r <: Exception

Exception thrown when an input for r is wrong.

Scats.internal.gen.ScatsGenWR_Δt — Type

ScatsGenWR_Δt <: Exception

Exception thrown when an input for Δt is wrong.

Scats.internal.gen.ScatsGenWR_α — Type

ScatsGenWR_α <: Exception

Exception thrown when an input for α is wrong.

Scats.internal.gen.ScatsGenWR_β — Type

ScatsGenWR_β <: Exception

Exception thrown when an input for β is wrong.

Scats.internal.gen.ScatsGenWR_γ — Type

ScatsGenWR_γ <: Exception

Exception thrown when an input for γ is wrong.

Scats.internal.gen.ScatsGenWR_ν — Type

ScatsGenWR_ν <: Exception

Exception thrown when an input for ν is wrong.

Scats.internal.gen.ScatsGenWR_ϕ — Type

ScatsGenWR_ϕ <: Exception

Exception thrown when an input for ϕ is wrong.

Scats.internal.gen.example — Method

example(file::AbstractString)

Generate an example of a file containing the generator parameters.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.gen.example(file)

Scats.internal.gen.read! — Method

read!(gen::GenStruct, file::AbstractString)

Read generator parameters from a file to an instance of GenStruct.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.gen.example(file)
s.gen.read!(file)

Scats.internal.gen.reset! — Method

reset!(gen::GenStruct)

Reset an instance of GenStruct to default values.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.gen.example(file)
s.gen.read!(file)
s.gen.reset!()