Input

Scats.internal.input — Module

Module containing a type for storage and interaction with input data.

Scats.internal.input.InputStruct — Type

InputStruct()

Instantiate this type to interact with input data.

Data

N::IT=0: sample size;

Δt::RT=0: sample step;

q::RT=0: significance level;

t::Vector{RT}=[]: time array;

x::Vector{RT}=[]: values array.

Methods

read!(file::AbstractString): read input data from a file;

write(file::AbstractString): write input data to a file;

example(file::AbstractString): generate an example of the input/output file;

reset!(): reset an instance to default values.

Note

Data can be also read calling an instance like so:

using Scats
s = Scats.api()
file, _ = mktemp()
s.input.example(file)
s.input(file)

Scats.internal.input.ScatsInputEOF — Type

ScatsInputEOF <: Exception

Exception thrown when an end of file occurred.

Scats.internal.input.ScatsInputIsADir — Type

ScatsInputIsADir <: Exception

Exception thrown when the path is a directory.

Scats.internal.input.ScatsInputNotAFile — Type

ScatsInputNotAFile <: Exception

Exception thrown when the file is not found.

Scats.internal.input.ScatsInputWR_N — Type

ScatsInputWR_N <: Exception

Exception thrown when an input for N is wrong.

Scats.internal.input.ScatsInputWR_q — Type

ScatsInputWR_q <: Exception

Exception thrown when an input for q is wrong.

Scats.internal.input.ScatsInputWR_t — Type

ScatsInputWR_t <: Exception

Exception thrown when an input for t is wrong.

Scats.internal.input.ScatsInputWR_x — Type

ScatsInputWR_x <: Exception

Exception thrown when an input for x is wrong.

Scats.internal.input.ScatsInputWR_Δt — Type

ScatsInputWR_Δt <: Exception

Exception thrown when an input for Δt is wrong.

Scats.internal.input.example — Method

example(file::AbstractString)

Generate an example of the input/output file.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.input.example(file)

See: InputStruct.

Scats.internal.input.read! — Method

read!(input::InputStruct, file::AbstractString)

Read input data from a file to an instance of InputStruct.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.input.example(file)
s.input.read!(file)

Scats.internal.input.reset! — Method

reset!(input::InputStruct)

Reset an instance of InputStruct to default values.

Usage

using Scats
s = Scats.api()
s.input.reset!()

Scats.internal.input.write — Method

write(input::InputStruct, file::AbstractString)

Write input data from an instance of InputStruct to a file.

Usage

using Scats
s = Scats.api()
file, _ = mktemp()
s.input.write(file)