Public Interface

This part of the package is the only one supposed to be used by a regular user. All recommended for usage types and methods are accessible from an instance of Scats.api. For the inner parts of this package one should consider reading the pages in the Internals section.

Scats.api — Type

api()

Instantiate an instance of Scats API to get access to the public interface.

Types

input: input data;

gen: generator;

result: result data.

Methods

for input:

read_input!(file::AbstractString): read input data from a file;

write_input(file::AbstractString): write input data to a file;

input_example(file::AbstractString): generate an example of the input/output file.

Usage

using Scats
s = Scats.api()